ika_streamer/requirements.txt
aiofiles==0.6.0
beautifulsoup4==4.9.0
bleach==3.1.4
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
fastapi==0.54.1
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
h11==0.9.0
httplib2==0.17.3
httptools==0.1.1
idna==2.9
ikamail==0.0.6
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.5.1
pymongo==3.10.1
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
starlette==0.13.2
uritemplate==3.0.1
urllib3==1.25.9
uvicorn==0.11.3
uvloop==0.14.0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
kafka-python
genson
msgpack